change from int to size_t

+++ 14-binary_tree_balance.c
@@ -29,9 +29,9 @@ int binary_tree_balance(const binary_tree_t *tree)
  *
  * Return: Height of the tree, or 0 if tree is NULL.
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int left_height, right_height;
+	size_t left_height, right_height;
 
 	/* Check if tree is NULL */
 	if (tree == NULL)
